Assignment-1/Task-2/requirements.txt
flask==2.2.2
Werkzeug==2.3.7
pycryptodome==3.10.1
requests
